+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Select 'Acquisition method' filter from controlled vocabulary list. Refs UIPFPOL-33.
+
 ## [3.0.1](https://github.com/folio-org/ui-plugin-find-po-line/tree/v3.0.1) (2021-11-11)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-po-line/compare/v3.0.0...v3.0.1)